lahaina-qgki_defconfig: Use 12ms half-life for PELT
The half-life of a signal processed by the PELT algorithm refers to the median change point distance that the algorithm can detect. A smaller half-life means that the algorithm can detect more frequent changes in the signal.

Reducing the half-life of PELT to 12 ms may or may not be good depending on the specific application you are using it for. If you are working with signals that have very quick changes and you need to detect them accurately, then reducing the half-life to 12 ms may be helpful. However, if the changes in your signal are slower or less frequent, reducing the half-life may not be necessary and could potentially lead to overfitting (i.e., detecting false or insignificant change points).

So, whether or not reducing the half-life is good really depends on the specific use case and the characteristics of the signal you are working with.

Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>
Change-Id: I4478b0b8f60fd1d567920633496649bed3b608b9
Signed-off-by: TogoFire <togofire@mailfence.com>